+++ packages/collector/src/index.ts
@@ -6,7 +6,6 @@ export * from './consent';
 export * from './flow';
 export * from './push';
 export * from './destination';
-export * from './destination-code';
 export * from './handle';
 export * from './on';
 export * from './source';