+++ packages/server/destinations/gcp/src/index.ts
@@ -1,3 +1,5 @@
 // Google BigQuery
 export { destinationBigQuery } from './bigquery';
 export * as DestinationBigQuery from './bigquery/types';
+
+export { destinationBigQuery as default } from './bigquery';