+++ packages/server/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @walkeros/server-core
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [f39d9fb]
+- Updated dependencies [888bbdf]
+  - @walkeros/core@1.2.0
+
 ## 1.0.1
 
 ### Patch Changes