+++ packages/server/transformers/fingerprint/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @walkeros/server-transformer-fingerprint
 
+## 6.0.0
+
+### Patch Changes
+
+- 6778ab2: Add default exports for simpler CLI flow.json configuration
+- Updated dependencies [f39d9fb]
+- Updated dependencies [888bbdf]
+  - @walkeros/core@1.2.0
+  - @walkeros/server-core@1.0.2
+
 ## 5.0.0
 
 ### Patch Changes