+++ packages/web/destinations/api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @walkeros/web-destination-api
 
+## 1.1.2
+
+### Patch Changes
+
+- @walkeros/web-core@1.0.2
+
 ## 1.1.1
 
 ### Patch Changes