+++ packages/web/destinations/meta/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @walkeros/web-destination-meta
 
+## 1.0.2
+
+### Patch Changes
+
+- @walkeros/web-core@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes